+++ Entities/PEmitter.cpp
@@ -54,6 +54,7 @@ namespace RTE {
 		m_AvgBurstImpulse = -1.0f;
 		m_AvgImpulse = -1.0f;
 		m_FlashOnlyOnBurst = true;
+		m_SustainBurstSound = false;
 		m_LoudnessOnEmit = 1.0f;
 	}
 
@@ -101,6 +102,7 @@ namespace RTE {
 		m_EmissionOffset = reference.m_EmissionOffset;
 		m_FlashScale = reference.m_FlashScale;
 		m_FlashOnlyOnBurst = reference.m_FlashOnlyOnBurst;
+		m_SustainBurstSound = reference.m_SustainBurstSound;
 		m_LoudnessOnEmit = reference.m_LoudnessOnEmit;
 
 		return 0;
@@ -158,6 +160,7 @@ namespace RTE {
 		MatchProperty("EmissionOffset", { reader >> m_EmissionOffset; });
 		MatchProperty("FlashScale", { reader >> m_FlashScale; });
 		MatchProperty("FlashOnlyOnBurst", { reader >> m_FlashOnlyOnBurst; });
+		MatchProperty("SustainBurstSound", { reader >> m_SustainBurstSound; });
 		MatchProperty("LoudnessOnEmit", { reader >> m_LoudnessOnEmit; });
 		
 		EndPropertyList;
@@ -213,6 +216,8 @@ namespace RTE {
 		writer << m_FlashScale;
 		writer.NewProperty("FlashOnlyOnBurst");
 		writer << m_FlashOnlyOnBurst;
+		writer.NewProperty("SustainBurstSound");
+		writer << m_SustainBurstSound;
 		writer.NewProperty("LoudnessOnEmit");
 		writer << m_LoudnessOnEmit;
 
@@ -485,6 +490,7 @@ namespace RTE {
 		else if (m_WasEmitting)
 		{
 			m_EmissionSound.Stop();
+			if (!m_SustainBurstSound) { m_BurstSound.Stop(); }
 			m_EndSound.Play(m_Pos);
 		}
 	}