+++ CHANGELOG.md
@@ -64,6 +64,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 	`JustStartedEmitting()`, which returns whether the emitter just started emitting this frame.  
 	`WasEmitting()`, which returns whether the emitter was emitting last frame.  
 
+- New `AEmitter` and `PEmitter` INI property `SustainBurstSound`, which determines whether the burst sound should sustain and play until completion even when the emitter stops emitting, or if it cuts out with the emitter. Defaults to false.  
+
 </details>
 
 <details><summary><b>Changed</b></summary>
@@ -117,8 +119,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixed a bug where AI pathfinding would choose unoptimal routes.
 
 - Fixed a crash that could occur when scripted objects were deleted, particularly when switching scene.
-
-- Fixed a bug where an emitter's burst sound would cut out as soon as the emitter stopped firing.
 
 </details>